+++ app/src/main/java/com/cashmanagment/models/CounterModel.kt
@@ -2,5 +2,5 @@ package com.cashmanagment.models
 
 data class CounterModel (
     val id: Int,
-    val count: Double
+    val count: Int
 )
