+++ app/src/main/java/com/cashmanagment/models/HistoryModel.kt
@@ -5,5 +5,5 @@ data class HistoryModel (
     val type: String,
     val description: String,
     val tag: String,
-    val amount : Double
+    val amount : Int
 )
